### 2019-08-15 #1649 @danielskatz
👋 @ezsolti - please merge https://github.com/ezsolti/feign/pull/6 and then enter a new comment here: ```@whedon generate pdf``` - if there are additional problems, please make changes to the source and regenerate the pdf until you are happy with it.
