### 2019-08-15 #1649 @katyhuff
I can definitely take this one. Thanks @danielskatz .

@ezsolti : We can begin requesting reviewers when ezsolti/feign#6 is merged and when you are happy with the PDF that builds from `@whedon generate pdf` . I have some reviewers in mind, though your suggestions are certainly welcome.  If you have any suggestions for potential reviewers then please mention them here in this thread. [This list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission.
