+++ paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 The operator declarations of spent nuclear fuel assemblies are routinely verified for nuclear safeguards purposes to ensure its non-diversion and manipulation. Many countries consider the possibility of eventually placing the fuel in a geological repository and prior to this it is expected that the fuel assemblies need to be carefully characterized and verified, for instance using gamma spectroscopy measurements. It can be expected that in connection to such activities, verifying parameters such as burnup (ie. the energy outtake from the nuclear fuel), cooling time (ie. the time the fuel spent outside the reactor after operation), initial enrichment (ie. the amount of fissile material in the fuel before operation) and integrity (whether pins inside the assembly have been manipulated) may become an important task of nuclear safeguards inspectors, since discrepancies may indicate unauthorized activities at the facilities. Ideally, the verification should be done with non-destructive assay systems.
 
-Passive gamma spectroscopy provides a robust and relatively simple method to analyze spent fuel since the characteristics of spent nuclear fuel strongly affect the gamma radiation emitted from the fuel [@Janssonthesis]. Lately, passive gamma tomography has also become a possible method to characterize and analyse properties of spent nuclear fuels [@pget]. In both cases, gamma radiation is measured around the fuel assembly from a distance using one or more collimated detectors with spectroscopic capabilities. The detector efficiency, the ratio between number of detected particles and number of particles emitted by the source, of such systems is of great interest. The detector efficiency is the product of the geometric efficiency (probability that emitted particles reach the detector region) and the intrinsic efficiency of the detector (probability that the particles are detected). The detector efficiency is usually estimated with Monte Carlo simulations, most often by using the MCNP particle transport code [@mcnp]. Nevertheless, analog Monte Carlo simulations of spent fuel passive gamma measurements are extremely time consuming. This depends mainly on the facts that the source (the spent fuel assembly) is both highly radioactive and highly attenuating (made of high-density uranium dioxide), and that the strong radiation field around the source often makes it difficult to place it near the detectors. In addition, it is not uncommon to use collimators with narrow slits around the detectors, something that makes the simulation of the detection process even more time consuming. Thus, only a tiny fraction of source particles reach the detector. As a solution, the simulation is often split into two parts: first, a point-detector (F5 in the MCNP jargon) tally is used to estimate the energy-resolved gamma photon flux at the location of the detector, and then a subsequent pulse-height tally computation (F8 in the MCNP jargon) is done to estimate the detector response. The F5 tally provides a semi-deterministic solution of the transport problem. The history of the gamma-ray is tracked with Monte Carlo, however at each interaction of the particle, the contribution to a point-detector is calculated analytically. A great advantage of this method is that both the direct attenuation of gamma radiation and the build-up factor is included. However, this methodology still requires excessive amount of computing time when the efficiency curves of a large number of assembly configurations are needed, for example when training machine learning algorithms for classification analysis of intact and manipulated fuel assemblies based on the measured gamma spectrum [@elterphysor]. Furthermore, when prototyping with such algorithms, the accuracy of a Monte Carlo code may not be needed. Nevertheless, in collimated passive gamma spectroscopy setups, one could argue, that only the uncollided gamma-rays contribute significantly to the gamma peaks in the spectrum (if the background is removed), thus computing the uncollided F5 tally would be satisfactory. However, computing the uncollided point-detector flux does not require a transport code such as MCNP, the task can be solved with a program that feigns to be a transport code such as the one developed here.
[…]
