### 2019-08-15 afb90a4
inline math fixed
