+++ paper/paper.md
@@ -35,11 +35,11 @@ Passive gamma spectroscopy provides a robust and relatively simple method to ana
 
 ``feign`` is a Python package which implements a 2D point-kernel method to estimate the uncollided point-detector gamma flux around a rectangular spent fuel assembly. The user defines the experimental setup: materials, pin types (consisting of nested annular material regions), assembly lattice, detector locations and optionally collimators and additional absorber elements. Then, the program iterates through each lattice position containing source material and calculates the distance traveled in various materials towards the detector point by a gamma-ray emitted from the given lattice position. Based on the user input, the actual source location is either the center of the pin or a randomly selected location inside the pin (the user then also sets the number of randomly selected locations per pin, in order to estimate the standard deviation of the traveled distances). 
 
-A traveled distance map can be seen in Fig. \ref{fig:distance} for a 17x17 pressurized water reactor (PWR) fuel assembly measured at the passive gamma spectroscopy station of the Swedish Central Interim Storage Facility for Spent Nuclear Fuel (Clab) as described by [@Vaccaro2016]. The detector point is facing the upper right corner. Based on the traveled distance maps and user-provided mass attenuation coefficient data, the program evaluates the probabilities $$P_i(E)$$ (or point-flux per source particle)
+A traveled distance map can be seen in Fig. \ref{fig:distance} for a 17x17 pressurized water reactor (PWR) fuel assembly measured at the passive gamma spectroscopy station of the Swedish Central Interim Storage Facility for Spent Nuclear Fuel (Clab) as described by [@Vaccaro2016]. The detector point is facing the upper right corner. Based on the traveled distance maps and user-provided mass attenuation coefficient data, the program evaluates the probabilities $P_i(E)$ (or point-flux per source particle)
 
 $$P_i(E) =\frac{1}{4\pi R_{i}^2}\prod\limits_m e^{-\mu_m(E) d_{i,m}}$$
 
-for each energy _E_ requested by the user that a gamma-ray emitted from position _i_ hits the detector without collision. $$R_i$$ is the distance between the source position and the detector, $$\mu_m$$ is the total or average attenuation coefficient of material _m_ and $$d_{i,m}$$ is the distance traveled by a gamma-ray emitted from position _i_ through material _m_. A contribution map can be seen in Fig. \ref{fig:contribution} for the same 17x17 assembly. When summing up the contributions made by each pin for each energy, one gets the geometric efficiency of the system. Fig. \ref{fig:geomeff} illustrates one case when the source locations are the center of the pins and another case when ten source locations were randomly selected in each pin. One can see that the pin center approximation has a lower efficiency as compared to the case with randomly selected source locations inside the pin. This is due to that the pin center case overestimates the traveled distances (thus underestimate the geometric efficiency). 
+for each energy $E$ requested by the user that a gamma-ray emitted from position $i$ hits the detector without collision. $R_i$ is the distance between the source position and the detector, $\mu_m$ is the total attenuation coefficient of material $m$ and $d_{i,m}$ is the distance traveled by a gamma-ray emitted from position $i$ through material $m$. A contribution map can be seen in Fig. \ref{fig:contribution} for the same 17x17 assembly. When summing up the contributions made by each pin for each energy, one gets the geometric efficiency curve of the system. Fig. \ref{fig:geomeff} illustrates one case when the source locations are the center of the pins and another case when ten source locations were randomly selected in each pin. One can see that the pin center approximation has a lower efficiency as compared to the case with randomly selected source locations inside the pin. This is due to that the pin center case overestimates the traveled distances (thus underestimate the geometric efficiency). 
 
[…]
