+++ paper/paper.bib
@@ -59,7 +59,7 @@ doi = {10.13182/NT11-135}}
 
 @Conference{elterMVA,
   Title                    = {Partial defect identification in PWR spent fuel using Passive Gamma Spectroscopy},
-  Author                   = {Zs. Elter, L. Caldeira Balkest\r{a}hl, S. Grape, C. Hellesen},
+  Author                   = {Zs. Elter, L. Caldeira Balkeståhl, S. Grape, C. Hellesen},
   Booktitle                = {Proceedings of the PHYSOR 2018, Cancun, Mexico},
   Year                     = {2018},
 
