+++ paper/paper.bib
@@ -58,8 +58,8 @@ doi = {10.13182/NT11-135}}
 }
 
 @Conference{elterMVA,
-  Title                    = {Partial defect identification in PWR spent fuel using Passive Gamma Spectroscopy},
-  Author                   = {Zs. Elter, L. Caldeira Balkeståhl, S. Grape, C. Hellesen},
+  Title                    = {Partial defect identification in {PWR} spent fuel using Passive Gamma Spectroscopy},
+  Author                   = {Zs. Elter and L. Caldeira Balkeståhl and S. Grape and C. Hellesen},
   Booktitle                = {Proceedings of the PHYSOR 2018, Cancun, Mexico},
   Year                     = {2018},
 
