### 2019-08-15 #1649 @ezsolti
@katyhuff : Hello! Thank you for taking care of the paper. I reached a pdf what I'm more or less satisfied with (besides that a cite to a conference paper doesn't want to render correctly, what ever I do).
I have reviewers in mind only from the safeguards side, so maybe it is better to go with your suggestions.
