### 2019-08-15 #1649 @katyhuff
@ezsolti :

Ok --  I'll start asking for reviewers. (nonproliferation and safeguards suggestions are just as welcome as anyone else).

How would you like the citation to render?
