### 2019-08-15 #1649 @ezsolti
@katyhuff : In that case Stefano Vaccaro (author of one of the cited papers: https://www.sciencedirect.com/science/article/pii/S0168900216307707) might be a good candidate.

For the 'Partial defect identification in PWR spent fuel using Passive Gamma Spectroscopy' paper the Proceedings doesn't want to appear, also when referenced all surnames appear instead of (Elter et al., 2018).
