### 2019-08-15 #1649 @katyhuff
* @sskutnik: Prof. Skutnik, are you able to and interested in taking on this review? The Journal of Open Source Software is a developer friendly, open access journal for research software packages. A JOSS review primarily involves checking submissions against a checklist of essential software features and details in the submitted paper and code. I am the handling editor for a submission that could use your expertise. More information about the review process can be found in [our reviewer documentation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). If you are willing and able to review this submission for JOSS, please let me know in this github thread and fill out your information at this signup form: https://joss.theoj.org/reviewer-signup.html . If not, please just let me know in this thread.
* This submission is in need of expertise in nuclear engineering, particularly gamma spectroscopy and ray tracing, as well as familiarity with the systems and languages involved (python).
* Title: feign: a Python package to estimate geometric efficiency in passive gamma spectroscopy measurements of nuclear fuel
* Summary: feign is a python package for estimating the geometric efficiency in passive gamma spectroscopy measurements of spent nuclear fuel assemblies. It implements a 2D point-kernel method without build-up factors (ie. an "uncollided F5 tally" as known by MCNP users). The name feign implies that the program pretends to be a transport code, however it is rather a ray-tracing code. It is intended for nuclear safeguards specialists and nuclear engineers who want to get a quick estimate on the geometric efficiency in their passive gamma setup. It might be also useful to people working with passive gamma emission tomography of spent fuel. The feign API allows the user to define the geometry of a rectangular fuel assembly, which is built of pins (nested annular material regions). The user also defines the composition of materials present in the simulation, the detector points where the efficiency needs to be evaluated, and optionally collimators and absorber elements.
* Article Proof: https://github.com/openjournals/joss-papers/blob/joss.01649/joss.01649/10.21105.joss.01649.pdf
* Submitting author: @ezsolti (Zsolt Elter)
* Repository: https://github.com/ezsolti/feign
* Version: v1.0.0
