### 2019-08-15 #1650 @whedon
[REVIEW]: feign: a Python package to estimate geometric efficiency in passive gamma spectroscopy measurements of nuclear fuel
**Submitting author:** @ezsolti (<a href="http://orcid.org/0000-0003-2339-4340">Zsolt Elter</a>)
**Repository:** <a href="https://github.com/ezsolti/feign/" target ="_blank">https://github.com/ezsolti/feign/</a>
**Version:** v1.0.0
