### 2019-08-15 #1649 @katyhuff
Ok, folks, let's move over to #1650 to conduct the review. In that issue, each of you have a checklist. Please make sure that you have notifications turned on in issue #1650 and that you have at some point completed the reviewer signup form. https://joss.theoj.org/reviewer-signup.html
