+++ SerpentToXCOM/README.md
@@ -1,0 +1,6 @@
+# Serpent XCOM interface
+
+This folder contains some scripts to create mass attenuation coefficients
+from serpent bumat files with XCOM.
+
+The folder is under development, and is not a core part of ``feign``
