### 2019-08-16 678a2cb
1st commit to Serpent-XCOM helper interface. Under development
