+++ SerpentToXCOM/README.md
@@ -3,4 +3,4 @@
 This folder contains some scripts to create mass attenuation coefficients
 from serpent bumat files with XCOM.
 
-The folder is under development, and is not a core part of ``feign``
+The folder is under development, and is not a core part of ``feign``, hence it may be migrated to a separate repo when mature enough.
