### 2019-08-16 #1650 @katyhuff
@ezsolti : This issue regarding authorship can be helped with some rewriting of history, if you like...  https://help.github.com/en/articles/changing-author-info
