### 2019-08-16 #1650 @katyhuff
@kellyrowland @ezsolti : Since this will be the "v.1.0.0" release, it is not uncommon for authors to submit to JOSS before the release, and then, after making the edits requested by the reviewers (but not other major edits), make the v.1.0.0 release (since a release must be created for zenodo anyway, which is the last step of the JOSS publication.)

For now, let's assume this review process will end with a DOI that points to a v.1.0.0 release, but that all reviews are conducted at the most recent commit (at the time of submission: 61978385bd1b950ea66a9704572484dc20b93b34 ). @ezsolti This will require that, until acceptance and until after the release, you only make edits to your master branch that were requested by these two JOSS authors. Continuing development can happen in parallel of course, on an experimental branch -- but should not be included in the release, as it will not be included in the review.
