### 2019-08-20 #1650 @ezsolti
Yes, thank you for the clarification @katyhuff. I am not supposed to change the code (which i havent done), or anything in the repo, like minor typos in the the readme (which i have done, but can go back to the original commit)?
