### 2019-08-20 #1650 @katyhuff
@ezsolti : very minor changes are fine, as are changes related to satisfying the recommendations of the reviewers. It's just that, for cleanliness of the meaning of the review process, major changes shouldn't be incorporated mid-review (unless called for by the review process itself). Don't worry, we try to keep it really quick, and you can always work in another branch to keep your master branch clean.
