### 2019-08-29 ezsolti/feign#7 @sskutnik
JOSS paper comments
## Figure 1 caption

Would be useful to indicate the detector location (upper right corner) within the caption for Figure 1. This is in the text, but it'd be helpful to be within the caption as well.

## Determining relative efficiency from pin center vs. random points

Specifically with respect to this text:
> Fig. \ref{fig:geomeff} illustrates one case when the source locations are the center of the pins and another case when ten source locations were randomly selected in each pin. One can see that the pin center approximation has a lower efficiency as compared to the case with randomly selected source locations inside the pin. **This is due to that the pin center case overestimates the traveled distances (thus underestimate the geometric efficiency).**

1. Why only select ten source locations for this evaluation?
2. How does the pin center over-estimate the traveled distance? If anything, it seems like this would be the mean value for the distribution of points throughout the pin. (i.e., as the pin is a circle, if we assume emissions are uniform throughout the pin, the integral probability for 0 < theta < pi is equal to that of pi < theta < 2 pi). Is there any way to recover any data on the distribution of points within the pin used for sampling (or to simply sample more points)?
3. I might suggest some minor rephrasing for clarity, such as:
> This is due to the fact that assuming all emissions originate from the center of the pin overestimates the traveled distances (thus underestimating the geometric efficiency).

## References

- Reference 1 (``@elterMVA``) needs a DOI / locatable citation. Based on the BibTeX, I'd suggest maybe using the ``@inproceedings`` tag instead of ``@conference``; that should then render the conference title.
- Reference 3 (``@pget``) is the same; recommend using ``@inproceedings`` in order to render the title; please provide a URL or DOI if available.
- Reference 4 (``Janssonthesis``) provide a URL to the online document or DOI if feasible.

## Miscellaneous (non-blocking)

1. I might suggest saying "MCNP syntax" (or especially, "MCNP convention") instead of "MCNP jargon"
2. The term "geometric efficiency" is a bit confusing given the more typical use to mean "solid angle subtended by the detector." I can see the validity of the use of this phrase here, but it may be useful to explicitly clarify that this is referring to the unattenuated contribution to a point detector.
