### 2019-08-29 #1650 @sskutnik
I've made a few minor comments re: the paper on ezsolti/feign/issues/7

In general, I'd recommend minor revisions, in part for clarity and also to clear up the theory of the efficiency calculation; also to update the references such that they render with fully-locatable citations. Otherwise, I think this looks fine.
