### 2019-08-29 #1650 @katyhuff
Thanks @sskutnik ! Excellent first review!

@ezsolti : I agree with Prof. Skutnik regarding clarity. When you feel you've handled these comments (in  ezsolti/feign/issues/7 mostly) please let me know (by @-mentioning me in this issue.) and we'll move forward.
