### 2019-09-03 d1288ab
paper and bib update according to review
