### 2019-09-03 ca6dd9e
update paper.md according to review
