### 2019-09-03 ab0d74e
update paper.bib according to review
