+++ paper/paper.bib
@@ -8,7 +8,8 @@
    ISSN = {1104-232X},
    number = {714},
    ISBN = {91-554-5315-5},
-   year = {2002}
+   year = {2002},
+   Url = {http://www.diva-portal.org/smash/record.jsf?pid=diva2%3A161704&dswid=8533}
 }
 
 
@@ -32,13 +33,15 @@ year  = {2012},
 publisher = {Taylor & Francis},
 doi = {10.13182/NT11-135}}
 
-@Conference{pget,
-  Title                    = {A Prototype for Passive Gamma Emission Tomography},
-  Author                   = {Honkamaa, Tapani and Levai, Ferenc and Berndt, R. and Schwalbach, Peter and Vaccaro, Stefano and Turunen, A.},
-  Booktitle                = {2014 Symposium on International Safeguards, At Vienna},
-  Year                     = {2014},
 
-}
+@INPROCEEDINGS{pget,
+author={M. {Mayorov} and T. {White} and A. {Lebrun} and J. {Brutscher} and J. {Keubler} and A. {Birnbaum} and V. {Ivanov} and T. {Honkamaa} and P. {Peura} and J. {Dahlberg}},
+booktitle={2017 IEEE Nuclear Science Symposium and Medical Imaging Conference (NSS/MIC)},
+title={Gamma Emission Tomography for the Inspection of Spent Nuclear Fuel},
+year={2017},
+doi={10.1109/NSSMIC.2017.8533017},
+ISSN={2577-0829},
+month={Oct},}
 
 
 @Article{Vaccaro2016,
@@ -49,20 +52,15 @@ doi = {10.13182/NT11-135}}
   Pages                    = {208 - 225},
   Volume                   = {833},
   Doi                      = {https://doi.org/10.1016/j.nima.2016.07.032},
-  File                     = {:home/zsolt/Documents/references/vaccaro1-s2.0-S0168900216307707-main(1).pdf:PDF},
   ISSN                     = {0168-9002},
-  Keywords                 = {Passive gamma},
-  Owner                    = {zsolt},
-  Timestamp                = {2017.10.26},
   Url                      = {https://www.sciencedirect.com/science/article/pii/S0168900216307707}
 }
 
-@Conference{elterMVA,
+@INPROCEEDINGS{elterMVA,
   Title                    = {Partial defect identification in {PWR} spent fuel using Passive Gamma Spectroscopy},
   Author                   = {Zs. Elter and L. Caldeira Balkeståhl and S. Grape and C. Hellesen},
   Booktitle                = {Proceedings of the PHYSOR 2018, Cancun, Mexico},
   Year                     = {2018},
 
 }
-
 
