### 2019-09-03 3a7e7a7
Update paper.bib according to review
