+++ paper/paper.bib
@@ -41,7 +41,7 @@ title={Gamma Emission Tomography for the Inspection of Spent Nuclear Fuel},
 year={2017},
 doi={10.1109/NSSMIC.2017.8533017},
 ISSN={2577-0829},
-month={Oct},}
+month={Oct}}
 
 
 @Article{Vaccaro2016,
@@ -58,9 +58,8 @@ month={Oct},}
 
 @INPROCEEDINGS{elterMVA,
   Title                    = {Partial defect identification in {PWR} spent fuel using Passive Gamma Spectroscopy},
-  Author                   = {Zs. Elter and L. Caldeira Balkeståhl and S. Grape and C. Hellesen},
-  Booktitle                = {Proceedings of the PHYSOR 2018, Cancun, Mexico},
-  Year                     = {2018},
-
+  Author                   = {{Zs}. {Elter} and L. {Caldeira Balkeståhl} and S. {Grape} and C. {Hellesen}},
+  Booktitle                = {Proceedings of PHYSOR 2018, Cancun, Mexico},
+  Year                     = {2018}
 }
 
