### 2019-09-03 ezsolti/feign#7 @ezsolti
@sskutnik : Thank you very much for the great comments. I modified the paper and the references based on your remarks. I have got a better reference for the PGET now. Regarding elterMVA, unfortunately the conference organizers still haven't put much effort into publishing the proceedings, so i cannot improve the citation to that one further. Shall I skip it? (Also, I still haven't mananged to render it correctly, maybe the Zs. messes things up, though I tried that also in various ways:))

Regarding your comment on the term "geometric efficiency". Would you propose to change the title as well? I actually had a dilemma about this term, whether I should go with it, or use something else. This one felt a bit more "obvious" for someone working with gamma spectroscopy, even if in this work only the unattenuated contribution is meant.

Related to the technical questions:

## Pin center vs random location

1, I selected only ten points because the pin diameter is reasonably small, and ten points seemed to sample the area well. Seemed to means here that I looked into how the mean curve converges with the number of random source sites, and it got converged already with ten points for the given geometry. Nevertheless, I thought this reasoning is slightly out of the scope of the paper. Do you think I should mention it?

2, Yes, the source locations are available in the [Experiment object](https://ezsolti.github.io/feign/experiment.html#feign.blocks.Experiment.sourcePoints). I have also felt it slightly counter-intuitive that the center case is not the same as the mean of the random locations, thus I have done an illustration for the simplest case. Let's consider only one pin (at (0,0), with radius 0.41cm), from which uncollided rays emitted from random locations are tracked to a detector point (at (170,170)). The random location is colored red when the distance traveled within the pin is less than the radius of the pin (which would be the center case), and blue when the distance traveled within the pin is greater than the radius.

![detsourcedist_distribution](https://user-images.githubusercontent.com/11595297/64196601-7f676c00-ce84-11e9-9b4c-d4d976608ab9.png)

The figure illustrates that on average rays travel a shorter distance within the pin than the radius of the pin. The magnitude of this discrepancy of course depends on the detector location. Also the presence of neighboring pins complicate things further. The quoted conclusion is thus not a general one, nevertheless for a detector placed far and facing the corner, the statement is valid, the traveled distance is greater for a source in the pin center than the average of random source locations. Similarly as in 1, I felt that such an explanation would be unnecessary for the paper, but I would gladly include it, if you feel it improves the article.

(Btw, having the opportunity to show this figure feels like when one can show a backup slide during a presentation:) So I just hope that I actually answered what you asked for:))
