### 2019-09-03 #1650 @ezsolti
@katyhuff : I have tried to address the comments of Prof. Skutnik, and as you see I have also tried to render a new paper based on some comments which did not require further "iterations" (I hope that this update was not against the rules). Unfortunately, one reference still renders strangely. Let's see whether @sskutnik would advise further changes based on the discussion in ezsolti/feign/issues/7 .
