### 2019-09-03 ezsolti/feign#7 @sskutnik
Hi @ezsolti, I think this does a great job of answering my question, thank you. The references look good; they are sufficiently specific as to be locatable (even if ANS conference proceedings aren't always as accessible as one would ideally like). I think the changes here are sufficient.

My personal feelings toward the figure would be toward including it in the paper if practical, if only because it supports what is a nominally counter-intuitive point about the relative attenuation factor; your explanation makes a good deal of sense, and I realize that this has to do with the nature of the point-detector problem now that I see your figure. (i.e., I think a ring detector or isotropic series of point detectors would not manifest this bias, but a point detector does. Since the application is for point detectors, I actually think this finding is fairly important to emphasize as to why the geometry calculation you've made is superior to using the center-pin approximation.) Basically, I think it only helps to support the case for why this software is useful.  (This may also be a good way to introduce the concept of the "experiment" object in your discussion of the software and how this is used.)

With respect to terminology, this is a tough one. In the reactor physics world, a similar concept is represented by the Dancoff Factor; this is not a precise analogy, but it's relatively similar. You might call it a "view factor" or even "effective attenuation length"; that said, I also respect the fact that this works much like geometric efficiency in gamma spectroscopy, so this is an issue I am willing to defer to the author's judgment on.
