### 2019-09-03 #1650 @sskutnik
I think I'm happy with the proposed changes; I'd leave it to @ezsolti whether to include the figure provided in https://github.com/ezsolti/feign/issues/7

Otherwise, I think all of my concerns have been addressed and I would happily recommend this to move forward for publication.
