### 2019-09-04 #1650 @ezsolti
@sskutnik : Thinking it over again, i might be on the opinion that the figure in ezsolti/feign#7 or a similar one for more pins should rather be included in the documentation when reasoning why using the random source case is superior to the center case in certain setups. Feels like it would overcomplicate the paper which has already reached the recommended upper limit for length.
