### 2019-09-04 #1650 @sskutnik
@ezsolti This seems like a reasonable conclusion. I think the documentation is a reasonable alternative.
