### 2019-09-21 #1650 @katyhuff
@ezsolti I have reviewed the code and the paper. I have some comments on the paper which can be found in the attached pdf:
[joss-feign.pdf](https://github.com/openjournals/joss-reviews/files/3638206/joss-feign.pdf)

Primarily:
- [ ]  In your references, the DOI https://doi.org/10.1016/j.nima.2016.07.032 is INVALID because of 'https://doi.org/' prefix
- [ ] The statement of need and the description of the intended usership of the work could be much more clear and should be stated earlier in the paper.
- [ ] There are a handful of grammatical errors (see pdf)
- [ ] The side-by-side figures would be improved by matching legend maxima so that the colors have the same meaning in adjacent images.
- [ ] The second paragraph is very long and quite heavy in jargon. The average JOSS reader will find it impenetrable. Consider more clear, jargon-free language.
- [ ] The paper should cite software on which feign relies (numpy, matplotlib)
