### 2019-09-25 cfcb569
Corrected grammar mistakes; updated figure+bib.
