+++ paper/paper.bib
@@ -13,15 +13,6 @@
 }
 
 
-@article{Jacobsson2017,
-author = {S. Jacobsson Sv\"{a}rd and L. Smith and T. White and V. Mozin and P. Jansson and P. Andersson and A. Davour and S. Grape and H. Trellue and N. Deshmukh and E. Miller and R. Wittman and T. Honkamaa and S. Vaccaro and J. Ely},
-title = {Outcomes of the {JNT} 1955 Phase {I} Viability Study of Gamma Emission Tomography for Spent Fuel Verification},
-journal = {ESARDA Bulletin},
-number = {55},
-year = {2017}
-
-}
-
 @article{mcnp,
 author = {T. Goorley and M. James and T. Booth and F. Brown and J. Bull and L. J. Cox and J. Durkee and J. Elson and M. Fensin and R. A. Forster and J. Hendricks and H. G. Hughes and R. Johns and B. Kiedrowski and R. Martz and S. Mashnik and G. McKinney and D. Pelowitz and R. Prael and J. Sweezy and L. Waters and T. Wilcox and T. Zukaitis},
 title = {Initial {MCNP6} Release Overview},
@@ -36,7 +27,7 @@ doi = {10.13182/NT11-135}}
 
 @INPROCEEDINGS{pget,
 author={M. {Mayorov} and T. {White} and A. {Lebrun} and J. {Brutscher} and J. {Keubler} and A. {Birnbaum} and V. {Ivanov} and T. {Honkamaa} and P. {Peura} and J. {Dahlberg}},
-booktitle={2017 IEEE Nuclear Science Symposium and Medical Imaging Conference (NSS/MIC)},
+booktitle={2017 {IEEE} Nuclear Science Symposium and Medical Imaging Conference ({NSS/MIC})},
 title={Gamma Emission Tomography for the Inspection of Spent Nuclear Fuel},
 year={2017},
 doi={10.1109/NSSMIC.2017.8533017},
@@ -51,15 +42,37 @@ month={Oct}}
   Year                     = {2016},
   Pages                    = {208 - 225},
   Volume                   = {833},
-  Doi                      = {https://doi.org/10.1016/j.nima.2016.07.032},
+  Doi                      = {10.1016/j.nima.2016.07.032},
   ISSN                     = {0168-9002},
   Url                      = {https://www.sciencedirect.com/science/article/pii/S0168900216307707}
 }
 
 @INPROCEEDINGS{elterMVA,
   Title                    = {Partial defect identification in {PWR} spent fuel using Passive Gamma Spectroscopy},
-  Author                   = {{Zs}. {Elter} and L. {Caldeira Balkeståhl} and S. {Grape} and C. {Hellesen}},
+  Author                   = {Elter, Zsolt and Caldeira Balkeståhl, Li and Grape, Sophie and Hellesen, Carl},
   Booktitle                = {Proceedings of PHYSOR 2018, Cancun, Mexico},
   Year                     = {2018}
+}
+
+@article{numpy,
+author = {Walt,Stéfan van der  and Colbert,S. Chris  and Varoquaux,Gaël },
+title = {The NumPy Array: A Structure for Efficient Numerical Computation},
+journal = {Computing in Science \& Engineering},
+volume = {13},
+number = {2},
+pages = {22-30},
+year = {2011},
+doi = {10.1109/MCSE.2011.37}
+}
+
+@article{matplotlib,
+author = {Hunter,John D. },
+title = {Matplotlib: A 2D Graphics Environment},
+journal = {Computing in Science \& Engineering},
+volume = {9},
+number = {3},
+pages = {90-95},
+year = {2007},
+doi = {10.1109/MCSE.2007.55},
 }
 
