+++ paper/paper.bib
@@ -73,6 +73,6 @@ volume = {9},
 number = {3},
 pages = {90-95},
 year = {2007},
-doi = {10.1109/MCSE.2007.55},
+doi = {10.1109/MCSE.2007.55}
 }
 
