+++ paper/paper.bib
@@ -27,7 +27,7 @@ doi = {10.13182/NT11-135}}
 
 @INPROCEEDINGS{pget,
 author={M. {Mayorov} and T. {White} and A. {Lebrun} and J. {Brutscher} and J. {Keubler} and A. {Birnbaum} and V. {Ivanov} and T. {Honkamaa} and P. {Peura} and J. {Dahlberg}},
-booktitle={2017 {IEEE} Nuclear Science Symposium and Medical Imaging Conference ({NSS/MIC})},
+booktitle={2017 {IEEE} {N}uclear {S}cience {S}ymposium and {M}edical {I}maging {C}onference ({NSS/MIC})},
 title={Gamma Emission Tomography for the Inspection of Spent Nuclear Fuel},
 year={2017},
 doi={10.1109/NSSMIC.2017.8533017},
@@ -50,13 +50,13 @@ month={Oct}}
 @INPROCEEDINGS{elterMVA,
   Title                    = {Partial defect identification in {PWR} spent fuel using Passive Gamma Spectroscopy},
   Author                   = {Elter, Zsolt and Caldeira Balkeståhl, Li and Grape, Sophie and Hellesen, Carl},
-  Booktitle                = {Proceedings of PHYSOR 2018, Cancun, Mexico},
+  Booktitle                = {Proceedings of {PHYSOR} 2018, {C}ancun, {M}exico},
   Year                     = {2018}
 }
 
 @article{numpy,
 author = {Walt,Stéfan van der  and Colbert,S. Chris  and Varoquaux,Gaël },
-title = {The NumPy Array: A Structure for Efficient Numerical Computation},
+title = {The {NumPy} Array: A Structure for Efficient Numerical Computation},
 journal = {Computing in Science \& Engineering},
 volume = {13},
 number = {2},
