+++ paper/paper.md
@@ -34,7 +34,7 @@ Passive gamma spectroscopy provides a robust and relatively simple method to ana
 
 # feign
 
-``feign`` is a Python package which implements a 2D point-kernel method to estimate the uncollided point-detector gamma flux around a rectangular spent fuel assembly. The package relies on NumPy [@numpy] for data handling and Matplotlib [@matplotlib] for plotting the geometry and the results. The user defines the experimental setup: materials, pin types (consisting of nested annular material regions), assembly lattice, detector locations and optionally collimators and additional absorber elements. Then, the program iterates through each lattice position containing source material and calculates the distance traveled in various materials towards the detector point by a gamma-ray emitted from the given lattice position. Based on the user input, the actual source location is either the center of the pin or a randomly selected location inside the pin (the user then also sets the number of randomly selected locations per pin, in order to estimate the standard deviation of the traveled distances). 
+``feign`` is a Python package which implements a 2D point-kernel method to estimate the uncollided point-detector gamma flux around a rectangular spent fuel assembly. It is intended for nuclear safeguards specialists and nuclear engineers who want to get a quick estimate on the geometric efficiency in their passive gamma setup. The package relies on NumPy [@numpy] for data handling and Matplotlib [@matplotlib] for plotting the geometry and the results. The user defines the experimental setup: materials, pin types (consisting of nested annular material regions), assembly lattice, detector locations and optionally collimators and additional absorber elements. Then, the program iterates through each lattice position containing source material and calculates the distance traveled in various materials towards the detector point by a gamma-ray emitted from the given lattice position. Based on the user input, the actual source location is either the center of the pin or a randomly selected location inside the pin (the user then also sets the number of randomly selected locations per pin, in order to estimate the standard deviation of the traveled distances). 
 
 A traveled distance map can be seen in Fig. \ref{fig:distance} for a 17x17 pressurized water reactor (PWR) fuel assembly measured at the passive gamma spectroscopy station of the Swedish Central Interim Storage Facility for Spent Nuclear Fuel (Clab) as described by [@Vaccaro2016]. The detector point is facing the upper right corner. Based on the traveled distance maps and user-provided mass attenuation coefficient data, the program evaluates the probabilities $P_i(E)$ (or point-flux per source particle)
 
