+++ paper/paper.md
@@ -40,7 +40,7 @@ A traveled distance map can be seen in Fig. \ref{fig:distance} for a 17x17 press
 
 $$P_i(E) =\frac{1}{4\pi R_{i}^2}\prod\limits_m e^{-\mu_m(E) d_{i,m}}$$
 
-that a gamma-ray emitted from position $i$ with energy $E$ hits the detector without collision. $R_i$ is the distance between the source position and the detector, $\mu_m$ is the total attenuation coefficient of material $m$ and $d_{i,m}$ is the distance traveled by a gamma-ray emitted from position $i$ through material $m$. A contribution map can be seen in Fig. \ref{fig:contribution} for the same 17x17 assembly. When summing up the contributions made by each pin for each energy, one gets the geometric efficiency curve of the system. Fig. \ref{fig:geomeff} illustrates one case when the source locations are the center of the pins and another case when ten source locations were randomly selected in each pin. One can see that the pin center approximation has a lower efficiency as compared to the case with randomly selected source locations inside the pin. This is due to the fact that assuming all emissions originate from the center of the pin overestimates the traveled distances (thus underestimating the geometric efficiency).
+that a gamma-ray emitted from position $i$ with energy $E$ hits the detector without collision. $R_i$ is the distance between the source position and the detector, $\mu_m$ is the total attenuation coefficient of material $m$ and $d_{i,m}$ is the distance traveled by a gamma-ray emitted from position $i$ through material $m$. A relative contribution map can be seen in Fig. \ref{fig:contribution} for the same 17x17 assembly. When summing up the contributions made by each pin for each energy, one gets the geometric efficiency curve of the system. Fig. \ref{fig:geomeff} illustrates one case when the source locations are the center of the pins and another case when ten source locations were randomly selected in each pin. One can see that the pin center approximation has a lower efficiency as compared to the case with randomly selected source locations inside the pin. This is due to the fact that assuming all emissions originate from the center of the pin overestimates the traveled distances (thus underestimating the geometric efficiency).
 
 ![Example of contributions made by a pin position to a detector facing the corner of a 17x17 PWR assembly at Clab. Each pixel represents the relative contribution that a gamma-ray emitted from that position directly hits the detector.\label{fig:contribution}](article_contribution.png)
 
