### 2019-09-25 #1650 @ezsolti
@katyhuff Thank you for your comments, I've managed to address most of them:

- I have removed the prefix
- I have clearly included the intended usership in the 2nd section.
- I modified the grammar mistakes. There was one mark in the review on the 2nd page ("if the background is removed"), what I was not sure about whether it was deliberate.
- I have tried before with the matching legend maxima, unfortunately it does not look nice, since there is an order of magnitude difference. Logarithmic scale also doesnt improve it visually. Alternatively, I have normalized with the sum of the pinwise contributions, and plotted a relative pinwise contribution in percentage. This one looks informative even when the range is the same on the colorbar. Please, let me know whether you agree with this change.
 - Regarding the second paragraph being heavy, I have a hard time here, since I grew to like that part, but I also understand if you feel it problematic for JOSS readers. I have to think how to modify it in a way that I don't feel I'm loosing information. If you mean specifically the F5 and F8 referencing, I could remove that though I have added it with the intention to serve as an extra clarification for possible nuclear readers.
- Similarly, I agree with you that it is a bit awkward that the feign section starts only after the rather lengthy summary. I have seen this format (ie. summary section, then package section) in several JOSS papers, and I liked it, but indeed, the intended usership gets a bit far from the beginning of the article. I will need some time to think about how to change it without making the paper worse:)
- I have included a sentence on what NumPy and Matplotlib is used for and added references. I'm not sure whether it smoothly fits there.
