### 2019-09-26 #1650 @katyhuff
@ezsolti Thank you for working toward addressing my comments.

Please let me know when you are able to:

- [ ] Reduce the jargon in the 2nd paragraph (indeed, throughout the paper, which is already unusually long for a JOSS paper). The F5 tally mention is not the only item that I find problematic from a jargon perspective. Please aim to make clear the high-level functionality and purpose of the software for a diverse, non-specialist audience.
- [ ] Make clear, much earlier and more succinctly, a Statement of Need that illustrates the research purpose of the software. (e.g.  what problems the software is designed to solve and who the target audience is).

Please note, these are not merely my suggestions, but are requirements at the core of our submission instructions, which can be found here:

https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain

![image](https://user-images.githubusercontent.com/393899/65655852-433aad80-dfe3-11e9-9ad4-ee080a149c57.png)
