### 2019-10-03 #1650 @ezsolti
Hello @katyhuff , after reading through couple of more JOSS articles, I saw that some authors directly include a Statement of Need sections, so I tried to go with that. I also dropped the jargon. Please let me know what do you think of the text now?
