### 2019-10-08 #1650 @katyhuff
Thank you @kellyrowland and @sskutnik for your reviews -- we couldn't do this without you.
Thank you @ezsolti for your submission and for engaging actively in the review process! I have looked over the paper, double-checked all the DOI links, and have conducted a high-level review of the code itself. Everything looks ship-shape to me. At this point, please double-check the paper yourself, review any lingering details in your code/readme/etc., and then make an archive of the reviewed software in Zenodo/figshare/other service. **Please be sure that the DOI metadata (title, authors, etc.) matches this JOSS submission.** If you need to update the version, that's fine, please just let me know if the version in the DOI is different from the version named in the submission. Once that's complete, please update this thread with the DOI of the archive, and I'll move forward with accepting the submission. Until then, now is your moment for final touchups!
