### 2019-10-11 #1650 @katyhuff
@openjournals/joss-eics I believe this is ready for acceptance. Over to you!
