### 2019-10-12 e719836
fix in-text citation, and small editorial fixes
For https://github.com/openjournals/joss-reviews/issues/1650
