### 2019-10-12 #1650 @labarba
@ezsolti — I'm the Associate Editor-in-Chief on rotation this week. I made a few edits on the paper via PR. Unfortunately the diff is not too helpful, because long paragraphs were written on the same source line. The top fix I wanted to enter was an in-text citation that didn't use the right [citation syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax ). I also removed a few unnecessary "Then" and changed "analyze" to US spelling, plus little things.
