### 2019-10-13 #1650 @ezsolti
@labarba Thank you! I have merged your commit.
