### 2019-10-13 #1650 @labarba
Huge thanks to our editor: @katyhuff, and reviewers: @kellyrowland, @sskutnik — your contributions to JOSS are greatly appreciated 🙏
